requirements.txt
requests>=2.31.0
fpdf2>=2.8.2
pypdf>=3.17.0
markdown2>=2.4.10
flask>=3.0.0
gunicorn>=21.2.0
python-dotenv>=1.0.0